Allow requesting only roles from the new request view
