+++ web/packages/teleterm/src/ui/DocumentAccessRequests/NewRequest/NewRequest.tsx
@@ -17,86 +17,35 @@
  */
 
 import styled from 'styled-components';
+import { Alert, Box, Link } from 'design';
 
-import { Alert, Box, Flex, Link, Indicator } from 'design';
-import { space, SpaceProps, width } from 'design/system';
-
-import {
-  ShowResources,
-  Cluster,
-} from 'gen-proto-ts/teleport/lib/teleterm/v1/cluster_pb';
-
-import { SearchPagination, SearchPanel } from 'shared/components/Search';
+import { ShowResources } from 'gen-proto-ts/teleport/lib/teleterm/v1/cluster_pb';
 import {
   ResourceList,
   ResourceMap,
 } from 'shared/components/AccessRequests/NewRequest';
 
-import {
-  PendingAccessRequest,
-  extractResourceRequestProperties,
-} from 'teleterm/ui/services/workspacesService/accessRequestsService';
+import { PendingAccessRequest } from 'teleterm/ui/services/workspacesService/accessRequestsService';
 
 import { useWorkspaceContext } from 'teleterm/ui/Documents';
 import { useAppContext } from 'teleterm/ui/appContextProvider';
 
-import useNewRequest, { ResourceKind } from './useNewRequest';
-
-const agentOptions: ResourceOption[] = [
-  { value: 'role', label: 'Roles' },
-  {
-    value: 'node',
-    label: 'Servers',
-  },
-  {
-    value: 'app',
-    label: 'Apps',
-  },
-  {
-    value: 'db',
-    label: 'Databases',
-  },
-  {
-    value: 'kube_cluster',
-    label: 'Kubes',
-  },
-];
-
 export function NewRequest() {
   const ctx = useAppContext();
 
-  const { rootClusterUri } = useWorkspaceContext();
-  const rootCluster = ctx.clustersService.findCluster(rootClusterUri);
-  if (rootCluster.showResources === ShowResources.UNSPECIFIED) {
-    return <Indicator />;
-  }
-  return <Inner rootCluster={rootCluster} />;
-}
-
-function Inner(props: { rootCluster: Cluster }) {
   const {
-    attempt,
-    agentFilter,
-    pageCount,
-    updateQuery,
-    updateSearch,
-    selectedResource,
-    customSort,
-    fetch,
-    fetchStatus,
-    onAgentLabelClick,
-    addedResources,
-    addOrRemoveResource,
-    updateResourceKind,
-    prevPage,
-    requestableRoles,
-    nextPage,
-    agents,
-    addedItemsCount,
-  } = useNewRequest(props.rootCluster);
-  const { documentsService, localClusterUri } = useWorkspaceContext();
+    rootClusterUri,
+    localClusterUri,
+    documentsService,
+    accessRequestsService,
+  } = useWorkspaceContext();
+  const rootCluster = ctx.clustersService.findCluster(rootClusterUri);
+  ctx.clustersService.useState();
 
-  const requestStarted = addedItemsCount > 0;
+  const loggedInUser = rootCluster?.loggedInUser;
+  const requestableRoles = loggedInUser?.requestableRoles || [];
+  const addedResources = accessRequestsService.getPendingAccessRequest();
+  const requestStarted = accessRequestsService.getAddedItemsCount() > 0;
 
   function openClusterDocument() {
     const doc = documentsService.createClusterDocument({
@@ -107,67 +56,27 @@ function Inner(props: { rootCluster: Cluster }) {
   }
 
   const doesUnifiedResourcesShowBothAccesibleAndRequestableResources =
-    props.rootCluster.showResources === ShowResources.REQUESTABLE;
-  // This means that we can only request roles.
-  // Let's hide all tabs in that case.
-  const filteredAgentOptions = isRequestingResourcesFromResourcesViewEnabled
-    ? []
-    : agentOptions;
-
-  const isRoleList = selectedResource === 'role';
+    rootCluster?.showResources === ShowResources.REQUESTABLE;
 
   return (
     <Layout mx="auto" px={5} pt={3} height="100%">
-      {attempt.status === 'failed' && (
-        <Alert
-          kind="danger"
-          details={attempt.statusText}
-          primaryAction={{ content: 'Retry', onClick: fetch }}
-        >
-          Could not load resources
-        </Alert>
-      )}
       <StyledMain>
-        <Flex mt={3} mb={3}>
-          {filteredAgentOptions.map(agent => (
-            <StyledNavButton
-              key={agent.value}
-              mr={6}
-              p={1}
-              active={selectedResource === agent.value}
-              onClick={() => updateResourceKind(agent.value)}
-            >
-              {agent.label}
-            </StyledNavButton>
-          ))}
-        </Flex>
-        {/* roles use client-side search */}
-        {!isRoleList && (
-          <SearchPanel
-            updateQuery={updateQuery}
-            updateSearch={updateSearch}
-            pageIndicators={pageCount}
-            filter={agentFilter}
-            disableSearch={fetchStatus === 'loading'}
-          />
-        )}
         <ResourceList
-          agents={agents}
-          selectedResource={selectedResource}
+          agents={[]}
+          selectedResource={'role'}
           requestStarted={requestStarted}
-          customSort={customSort}
-          onLabelClick={onAgentLabelClick}
+          customSort={undefined}
+          onLabelClick={() => {}}
           addedResources={toResourceMap(addedResources)}
-          addOrRemoveResource={addOrRemoveResource}
+          addOrRemoveResource={(kind, resourceId) => {
+            // We can only have roles here.
+            if (kind === 'role') {
+              accessRequestsService.addOrRemoveRole(resourceId);
+            }
+          }}
           requestableRoles={requestableRoles}
-          disableRows={fetchStatus === 'loading'}
+          disableRows={false}
         />
-        {!isRoleList && (
-          <SearchPagination
-            nextPage={fetchStatus === 'loading' ? null : nextPage}
-            prevPage={fetchStatus === 'loading' ? null : prevPage}
-          />
-        )}
       </StyledMain>
       <Alert kind="outline-info" mb={2}>
         {doesUnifiedResourcesShowBothAccesibleAndRequestableResources ? (
@@ -218,48 +127,11 @@ const Layout = styled(Box)`
   }
 `;
 
-interface StyledNavButtonProps extends SpaceProps {
-  active?: boolean;
-}
-
-const StyledNavButton = styled.button<StyledNavButtonProps>(props => {
-  return {
-    color: props.active
-      ? props.theme.colors.text.main
-      : props.theme.colors.text.slightlyMuted,
-    cursor: 'pointer',
-    display: 'inline-flex',
-    fontSize: '14px',
-    position: 'relative',
-    padding: '0',
-    marginRight: '24px',
-    textDecoration: 'none',
-    fontWeight: props.active ? 700 : 400,
-    outline: 'inherit',
-    border: 'none',
-    backgroundColor: 'inherit',
-    flexShrink: '0',
-    borderRadius: '4px',
-    fontFamily: 'inherit',
-
-    '&:hover, &:focus': {
-      background: props.theme.colors.spotBackground[0],
-    },
-    ...space(props),
-    ...width(props),
-  };
-});
-
 const StyledMain = styled.div`
   display: flex;
   flex-direction: column;
   flex: 1;
 `;
-
-type ResourceOption = {
-  value: ResourceKind;
-  label: string;
-};
 
 function toResourceMap(request: PendingAccessRequest): ResourceMap {
   const resourceMap: ResourceMap = {
@@ -279,12 +151,5 @@ function toResourceMap(request: PendingAccessRequest): ResourceMap {
     });
   }
 
-  if (request.kind === 'resource') {
-    request.resources.forEach(resourceRequest => {
-      const { kind, id, name } =
-        extractResourceRequestProperties(resourceRequest);
-      resourceMap[kind][id] = name;
-    });
-  }
   return resourceMap;
 }
