+++ web/packages/teleterm/src/ui/TopBar/AdditionalActions.tsx
@@ -22,8 +22,6 @@ import styled from 'styled-components';
 import { Flex, Text, Popover } from 'design';
 import * as icons from 'design/Icon';
 
-import { ShowResources } from 'gen-proto-ts/teleport/lib/teleterm/v1/cluster_pb';
-
 import { useAppContext } from 'teleterm/ui/appContextProvider';
 import { TopBarButton } from 'teleterm/ui/TopBar/TopBarButton';
 import { IAppContext } from 'teleterm/ui/types';
@@ -62,8 +60,6 @@ function useMenuItems(): MenuItem[] {
   const hasNoActiveWorkspace = !documentsService;
   const areAccessRequestsSupported =
     !!activeRootCluster?.features?.advancedAccessWorkflows;
-  const isRequestingResourcesFromResourcesViewEnabled =
-    activeRootCluster?.showResources === ShowResources.REQUESTABLE;
 
   const { platform } = mainProcessClient.getRuntimeSettings();
   const isDarwin = platform === 'darwin';
@@ -105,9 +101,7 @@ function useMenuItems(): MenuItem[] {
       },
     },
     {
-      title: isRequestingResourcesFromResourcesViewEnabled
-        ? 'New role request'
-        : 'New access request',
+      title: 'New role request',
       isVisible: areAccessRequestsSupported,
       prependSeparator: true,
       Icon: icons.Add,
@@ -115,9 +109,7 @@ function useMenuItems(): MenuItem[] {
         const doc = documentsService.createAccessRequestDocument({
           clusterUri: activeRootCluster.uri,
           state: 'creating',
-          title: isRequestingResourcesFromResourcesViewEnabled
-            ? 'New Role Request'
-            : 'New Access Request',
+          title: 'New Role Request',
         });
         documentsService.add(doc);
         documentsService.open(doc.uri);
